fix(cc): widen QUIC_MAX_DATAGRAM to u64 to avoid int multiplication overflow
The window-size macros multiply QUIC_MAX_DATAGRAM by a constant in int,
then widen to u64. Making the base literal u64 keeps the arithmetic in
u64 (bugprone-implicit-widening-of-multiplication-result).

+++ src/cc/cc.h
@@ -5,7 +5,7 @@
 
 /* RFC 9002 7: NewReno-style congestion control. Sizes in bytes. */
 
-#define QUIC_MAX_DATAGRAM 1200
+#define QUIC_MAX_DATAGRAM ((u64)1200)
 #define QUIC_CC_MIN_WINDOW (2 * QUIC_MAX_DATAGRAM)            /* kMinimumWindow */
 #define QUIC_CC_INIT_WINDOW (10 * QUIC_MAX_DATAGRAM)          /* kInitialWindow */
